Squashed commit of the following:
commit 9e6e2e9ed50b1c708985a0acbe281ceba3f4714e
Author: Demian Wassermann <demian.wassermann@inria.fr>
Date:   Sun Nov 27 19:53:34 2022 +0100

    removing different warnings

commit afbdb063bab35f212126b00ca2b5d8001fecf381
Author: Demian Wassermann <demian.wassermann@inria.fr>
Date:   Sun Nov 27 19:13:21 2022 +0100

    Optimise better the queries before executing

commit 65609ed299d01a05d92ae3ad22d2bc9fe1743459
Author: Demian Wassermann <demian.wassermann@inria.fr>
Date:   Sun Nov 27 19:06:21 2022 +0100

    fixing type issues

commit 702dabc9c119844f2f054f668b13dd454e11d7a8
Author: Demian Wassermann <demian.wassermann@inria.fr>
Date:   Sun Nov 27 18:46:52 2022 +0100

    Improve handling of set semantics in the backend

+++ neurolang/datalog/wrapped_collections.py
@@ -19,7 +19,7 @@ REBV = ReplaceExpressionsByValues(dict())
 
 class WrappedTypeMap:
     row_maps = {
-        np.int64: int,
+        np.integer: int,
         np.float64: float
     }
